leo5

+++ workspace/loja/src/main/java/models/PagamentoComBoleto.java
@@ -23,7 +23,7 @@ public class PagamentoComBoleto extends Pagamento{
 	//Constructors
 	public PagamentoComBoleto() {}
 	
-	public PagamentoComBoleto(Integer id, EstadoPagamento estado, Pedido pedido, Date dataVencimento, Date dataPagamento) {
+	public PagamentoComBoleto(Long id, EstadoPagamento estado, Pedido pedido, Date dataVencimento, Date dataPagamento) {
 		super(id, estado, pedido);
 		this.dataPagamento = dataPagamento;
 		this.dataVencimento = dataVencimento;
